feat(client): move files around, finish TODOs, refactor, cleanUp
