+++ apps/client/src/routes/no-found-page.tsx
@@ -1,0 +1,3 @@
+export default function NotFoundPage() {
+  return <div>not found</div>;
+}
