--- apps/client/src/routes/no-match.tsx
@@ -1,3 +1,0 @@
-export default function NoMatch() {
-  return <div>no-match</div>;
-}
