+++ apps/client/src/routes/protected/home/budget/budget-route.ts
@@ -2,7 +2,7 @@ import { lazyRouteComponent, Route } from '@tanstack/react-router';
 import * as z from 'zod';
 
 import { AppRouteError } from '@/components/app-route-error';
-import { homeLayoutRoute } from '@/routes/home/home-route';
+import { homeRoute } from '@/routes/protected/home/home-route';
 import { Months, months } from '@/types/app';
 
 const searchSchema = z.object({
@@ -17,9 +17,9 @@ const searchSchema = z.object({
 export type BudgetSearch = z.infer<typeof searchSchema>;
 
 export const budgetRoute = new Route({
-  getParentRoute: () => homeLayoutRoute,
+  getParentRoute: () => homeRoute,
   path: '/budget',
   validateSearch: searchSchema,
-  component: lazyRouteComponent(() => import('@/routes/home/budget/budget-page')),
+  component: lazyRouteComponent(() => import('@/routes/protected/home/budget/budget-page')),
   errorComponent: AppRouteError,
 });
