+++ apps/client/src/routes/protected/home/settings/settings-section-page.tsx
@@ -1,15 +1,13 @@
-import { useParams } from '@tanstack/react-router';
-
 import { AccountSection } from '@/app/settings/account/account-section';
 import { AppearanceSection } from '@/app/settings/appearance/appearance-section';
 import { CategoriesSection } from '@/app/settings/categories/categories-section';
 import { NotificationsSection } from '@/app/settings/notifications/notifications-section';
 import { ProfileSection } from '@/app/settings/profile/profile-section';
 import { Head } from '@/components/head';
-import { settingsSectionRoute } from '@/routes/home/settings/settings-route';
+import { settingsSectionRoute } from '@/routes/protected/home/settings/settings-section-route';
 
 export default function SettingsSectionPage() {
-  const params = useParams({ from: settingsSectionRoute.id });
+  const params = settingsSectionRoute.useParams();
 
   return (
     <>
