+++ apps/client/src/routes/protected/welcome/welcome-page.tsx
