+++ apps/client/src/utils/router.utils.ts
@@ -2,7 +2,7 @@ import { redirect } from '@tanstack/react-router';
 
 import { supabase } from '@/lib/supabase-client';
 
-export async function getSession(pathName: string) {
+export async function getSessionOrRedirect(pathName: string) {
   const session = await supabase.auth.getSession();
 
   if (!session.data.session?.access_token) {
